Add `Freezer.error_handler_spec` for saving errors
The structure of `Freezer.error_handler_spec` mimics that of
`Flask.error_handler_spec`:
    {blueprint_as_str: {status_code_as_int: url}}

If a user sets `Freezer.error_handler_spec` when freezing an
application, the URLs in the spec will be handled just like
the URLs from @freezer.register_generator functions, except
that the response objects will be checked to ensure that
they have the proper, non-200 status code.

If a URL is specified in both a register_generator function and in
the error_handler_spec, the register_generator function will win,
and the page will be tested for a 200 status code.