can yield a last_modified time for URLs
A fairly simple way of allowing incremental builds.

A generator can yield a tuple (url, values, last_modified), where
last_modified is a datetime.

If the corresponding output file already exists and is at least as new
as the last_modified time, it isn't rebuilt. If it doesn't exist, or
it's older than the last_modified time, it is rebuilt.