Remove labels
These badges were actually only drawn on GitHub, and they’re now pretty
useless.

The latest PyPI version is the same as the latest version displayed on the
project’s home page on GitHub. The build status is now in GitHub Actions, so
it’s only one click away from the home page.